refactor: move `collectResolvedThreadIds` to `recap.ts`
The function operates exclusively on `PreviousFinding[]`, a type owned by `recap.ts`. Co-locating it with its only input type clarifies module boundaries and lets `index.ts` import the helper from the same module as `previousFindings`. The unit tests follow the function to `recap.test.ts`.

Addresses manki review comment on PR #758
https://github.com/manki-review/manki/pull/758#discussion_r... (PRRT_kwDORvFD2s6DfSPL)
